Rename peeler/pyproject/parse.py to manifest_adapter.py
